# backend/docs/swagger.yaml
basePath: /
definitions:
...
  webhooks.WebhookReceivedResponse:
    properties:
      log_id:
        type: string
      message:
        type: string
    type: object
info:
  contact:
    email: admin@your-registry.example.com
    name: Registry Support
  description: Complete Terraform Module and Provider Registry with SCM integration,
    mirrors, and storage backends
  license:
    name: Apache-2.0
  termsOfService: https://your-registry.example.com/terms
  title: Terraform Registry API
  version: 1.0.0
paths:
...
  /webhooks/scm/{module_source_repo_id}/{secret}:
    post:
      consumes:
      - application/json
      description: |-
        Receives and processes incoming webhook events from SCM providers (GitHub, GitLab, Azure DevOps, Bitbucket).
        Two-layer security is applied: the URL-embedded secret (last path segment of the registered callback URL)
        is verified first with a constant-time comparison, and then the provider's HMAC payload signature is
        validated against the stored webhook secret. Both checks must pass before the payload is processed.
        Accepted events are logged. Tag-push events trigger asynchronous auto-publish when AutoPublish is enabled.
      parameters:
      - description: Module source repository link ID (UUID) — uniquely identifies
          the SCM-to-module mapping
        in: path
        name: module_source_repo_id
        required: true
        type: string
      - description: URL-embedded webhook secret generated at link time; used as a
          first-line constant-time guard before HMAC validation
        in: path
        name: secret
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/webhooks.WebhookReceivedResponse'
        "400":
          description: Invalid repository ID or malformed/unreadable payload
          schema:
            additionalProperties: true
            type: object
        "401":
          description: URL secret mismatch or HMAC payload signature invalid
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Repository link or SCM provider not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error (connector build, log write, etc.)
          schema:
            additionalProperties: true
            type: object
      summary: Receive SCM webhook
      tags:
      - Webhooks
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: '"JWT token or API key. For JWT: ''Bearer {token}''. For API Key:
      ''Bearer {api_key}''"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"